requirements.txt
tensorflow==2.6.0
numpy==1.19.5
pandas==1.3.3
matplotlib==3.4.3
scikit-learn==0.24.2
opencv-python==4.5.3.56
Pillow==8.3.2
h5py==3.1.0
streamlit==0.88.0